added gallery container component
