+++ src/components/Gallery/index.js
@@ -1,103 +1,31 @@
 // Gallery
-
-import './style.css';
-import { useState, useEffect } from 'react';
-import { useParams } from 'react-router-dom';
-import GalleryItem from './GalleryItem';
-
+import Button from "../Button";
+import Header from "../Header";
+import { useParams } from "react-router";
+import GalleryContainer from "./GalleryContainer";
+import './styles.css';
+import '../../Screens/styles.css'
 
 function Gallery() {
-  const [imgRefs, setImgRefs] = useState([]);
-  const [filteredImgRefs, setFilteredImgRefs] = useState([]);
-  const planetID = useParams();
-  const planet = planetID.planetID;
 
+  const params = useParams();
+  const planetID = params.planetID;
 
-  // api calls: Nasa Image and Video Library
-  useEffect(() => {
-
-    const planets = ['mars', 'uranus', 'saturn'];
-
-    const requests = planets.map((planet) => {
-      return fetch(`https://images-api.nasa.gov/search?q=${planet}&media_type=image`)
-        .then((response) => {
-          return response.json();
-        });
-    });
-
-    // creating a promise to return all the api call results at once
-    Promise.all(requests)
-      .then((jsonData) => {
-
-        const marsData = jsonData[0].collection.items;
-        const uranusData = jsonData[1].collection.items;
-        const saturnData = jsonData[2].collection.items;
-
-        // taking only the items we want from the call
-        const marsArray = [marsData[49], marsData[79], marsData[87]];
-        const uranusArray = [uranusData[1], uranusData[4], uranusData[7], uranusData[21]];
-        const saturnArray = [saturnData[5], saturnData[7], saturnData[8], saturnData[20]];
-
-
-        // adding a planet property to each array to allow us to filter through them later
-        const marsWithPlanet = marsArray.map((obj) => {
-          return { ...obj, planet: 'mars' };
-        });
-
-        const uranusWithPlanet = uranusArray.map((obj) => {
-          return { ...obj, planet: 'uranus' };
-        });
-
-        const saturnWithPlanet = saturnArray.map((obj) => {
-          return { ...obj, planet: 'saturn' };
-        });
-
-        // combining all into one array
-        const allWithPlanet = [...marsWithPlanet, ...uranusWithPlanet, ...saturnWithPlanet];
-
-        console.log(allWithPlanet)
-        setImgRefs(allWithPlanet);
-      })
-
-  }, [])
-
-
-
-  const getImages = (nameOfPlanet) => {
-    const copyOfImgRefs = [...imgRefs]
-
-    // conditional filtering
-    if (nameOfPlanet) {
-      const imgFiltered = copyOfImgRefs.filter((eachImg) => {
-        return eachImg.planet === nameOfPlanet;
-      });
-      setFilteredImgRefs(imgFiltered)
-    }
-  }
-
-
-  // this is just to test if it will work, will not remain on final
-  const handleClick = () => {
-    getImages(planet)
-  }
-
-
-
-
-  return (
-    <>
-      <button onClick={handleClick}>click</button>
-      <div className='gallery'>
-        {
-          filteredImgRefs.map(image => {
-            return (
-              <GalleryItem image={image} key={image.data[0].nasa_id} />
-            )
-          })
-        }
-      </div>
-    </>
-  )
+    return (
+        <>
+            <Header />
+            <section className="planetComponent">
+                <div>
+                    <div className="planetBox">
+                    <h1>{planetID}</h1>
+                    <Button intent="closeButton" />
+                    </div>
+                    <GalleryContainer />
+                    <Button text="See Dates" />
+                </div>
+            </section>
+        </>
+    );
 }
 
 export default Gallery;
