+++ src/components/Gallery/styles.css
@@ -8,7 +8,3 @@ img{
   width: 100%;
   object-fit: fill;
 }
-
-/* .galleryImage{
-
-} */
